Add files via upload
This analysis highlights key concepts in data analytics, focusing on how centrality and dispersion measures can be applied to real-world datasets. Using the variable AGE from a dating profiles dataset, measures like median and interquartile range (IQR) are calculated and explored through jackknife and bootstrap estimators. The analysis reveals that both the median and IQR are unbiased with minimal error, making them robust metrics in the presence of outliers.

In addition, the relationship between airfare and the number of passengers is examined using linear regression techniques. The study explores transformations, such as logarithmic and square-root, to assess the impact on model assumptions like linearity, normality, and homoscedasticity. While some violations of these assumptions persist, the analysis provides valuable insights into the factors influencing airfare, particularly the inverse relationship with the number of passengers.